Unify TimeAxisBottom and AxisBottom
