+++ packages/ui-components/src/components/Axis/index.ts
@@ -1,3 +1,2 @@
 export * from "./AxisLeft";
 export * from "./AxisBottom";
-export * from "./TimeAxisBottom";
